requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2024.2.2
charset-normalizer==3.3.2
filelock==3.12.3
Flask==2.3.3
Flask-Cors==4.0.0
frozenlist==1.4.0
fsspec==2023.9.0
h5py==3.9.0
httplib2==0.22.0
idna==3.6
joblib==1.3.2
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.2
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.28.2
schedule==1.2.1
soupsieve==2.5
sympy==1.12
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tqdm==4.66.1
urllib3==1.26.18
Werkzeug==2.3.7
